+++ providers/ObjectScriptCompletionItemProvider.ts
@@ -5,6 +5,7 @@ import systemFunctions = require('./completion/systemFunctions.json');
 import systemVariables = require('./completion/systemVariables.json');
 import structuredSystemVariables = require('./completion/structuredSystemVariables.json');
 import { ClassDefinition } from '../utils/classDefinition.js';
+import { currentFile } from '../utils/index.js';
 
 export class ObjectScriptCompletionItemProvider implements vscode.CompletionItemProvider {
   provideCompletionItems(
@@ -193,23 +194,40 @@ export class ObjectScriptCompletionItemProvider implements vscode.CompletionItem
   ): vscode.ProviderResult<vscode.CompletionItem[] | vscode.CompletionList> {
     let range = document.getWordRangeAtPosition(position, /%?\b\w+[\w\d]*\b/) || new vscode.Range(position, position);
     let textBefore = document.getText(new vscode.Range(new vscode.Position(position.line, 0), range.start));
+    let curFile = currentFile();
+
+    const method = el => ({
+      label: el.name,
+      documentation: el.desc.length ? new vscode.MarkdownString(el.desc.join('')) : null,
+      kind: vscode.CompletionItemKind.Method,
+      insertText: new vscode.SnippetString(`${el.name}($0)`)
+    });
+
+    const property = el => ({
+      label: el.name,
+      documentation: el.desc.length ? new vscode.MarkdownString(el.desc.join('')) : null,
+      kind: vscode.CompletionItemKind.Property,
+      insertText: new vscode.SnippetString(`${el.name}`)
+    });
 
     let classRef = textBefore.match(/##class\(([^)]+)\)\.$/i);
     if (classRef) {
       let [, className] = classRef;
       let classDef = new ClassDefinition(className);
-      return classDef.methods('class').then(methods => {
-        let completion = [
-          ...methods.map(el => ({
-            label: el.name,
-            documentation: el.desc.length ? new vscode.MarkdownString(el.desc.join('')) : null,
-            kind: vscode.CompletionItemKind.Method,
-            insertText: new vscode.SnippetString(`${el.name}($0)`)
-          }))
-        ];
-        return completion;
-      });
+      return classDef.methods('class').then(data => data.map(method));
     }
+
+    if (curFile.fileName.endsWith('cls')) {
+      let selfRef = textBefore.match(/(?<!\.)\.\.$/i);
+      if (selfRef) {
+        let classDef = new ClassDefinition(curFile.name);
+        return Promise.all([classDef.methods(), classDef.properties()]).then(data => {
+          let [methods, properties] = data;
+          return [...methods.map(method), ...properties.map(property)];
+        });
+      }
+    }
+
     return null;
   }
 }