+++ utils/classDefinition.ts
@@ -1,4 +1,7 @@
+import * as vscode from 'vscode';
 import { AtelierAPI } from '../api';
+import { onlyUnique } from '.';
+import { DocumentContentProvider } from '../providers/DocumentContentProvider';
 
 export class ClassDefinition {
   private _className: string;
@@ -15,7 +18,7 @@ export class ClassDefinition {
     this._classFileName = ClassDefinition.normalizeClassName(className, true);
   }
 
-  async methods(scope: 'any' | 'class' | 'instance'): Promise<any[]> {
+  async methods(scope: 'any' | 'class' | 'instance' = 'any'): Promise<any[]> {
     let methods = [];
     let filterScope = method => scope === 'any' || method.scope === scope;
     const api = new AtelierAPI();
@@ -28,23 +31,43 @@ export class ClassDefinition {
       if (extend.length) {
         return api.actionIndex(extend).then(data => getMethods(data.result.content));
       }
-      return methods.filter(filterScope);
+      return methods
+        .filter(filterScope)
+        .filter(onlyUnique)
+        .sort();
     };
     return api.actionIndex([this._classFileName]).then(data => getMethods(data.result.content));
+  }
+
+  async properties(): Promise<any[]> {
+    let properties = [];
+    const api = new AtelierAPI();
+    const getProperties = content => {
+      let extend = [];
+      content.forEach(el => {
+        properties.push(...el.content.properties);
+        extend.push(...el.content.super.map(extendName => ClassDefinition.normalizeClassName(extendName, true)));
+      });
+      if (extend.length) {
+        return api.actionIndex(extend).then(data => getProperties(data.result.content));
+      }
+      return properties.filter(onlyUnique).sort();
+    };
+    return api.actionIndex([this._classFileName]).then(data => getProperties(data.result.content));
   }
 
   async super(): Promise<string[]> {
     const api = new AtelierAPI();
     let sql = `SELECT PrimarySuper FROM %Dictionary.CompiledClass WHERE Name = ?`;
-    return api
-      .actionQuery(sql, [this._className])
-      .then(data =>
+    return api.actionQuery(sql, [this._className]).then(
+      data =>
         data.result.content.reduce(
           (list: string[], el: { PrimarySuper: string }) =>
             list.concat(el.PrimarySuper.split('~').filter(el => el.length)),
           []
         )
-      );
+      // .filter(name => !['%Library.Base', '%Library.SystemBase'].includes(name))
+    );
   }
 
   async includeCode(): Promise<string[]> {
@@ -60,5 +83,50 @@ export class ClassDefinition {
           defaultIncludes
         )
       );
+  }
+
+  async getPosition(name: string, document: vscode.TextDocument): Promise<vscode.Location[]> {
+    let pattern = `((Class)?Method|Property|RelationShip) ${name}(?!\w)`;
+    let foundLine;
+    if (document) {
+      for (let i = 0; i < document.lineCount; i++) {
+        let line = document.lineAt(i);
+        if (line.text.match(pattern)) {
+          foundLine = i;
+          break;
+        }
+      }
+    }
+    let result: vscode.Location[] = [];
+    if (foundLine) {
+      result.push({
+        uri: DocumentContentProvider.getUri(this._classFileName),
+        range: new vscode.Range(foundLine, 0, foundLine, 0)
+      });
+    }
+    let extendList = await this.super();
+    let api = new AtelierAPI();
+    let docs = [];
+    extendList.forEach(async docName => {
+      docName = ClassDefinition.normalizeClassName(docName, true);
+      docs.push(api.getDoc(docName));
+    });
+    return Promise.all(docs).then((docs: any[]) => {
+      for (let doc of docs) {
+        if (doc && doc.result.content) {
+          let docName = doc.result.name;
+          let content = doc.result.content;
+          for (let line of content.keys()) {
+            if (content[line].match(pattern)) {
+              result.push({
+                uri: DocumentContentProvider.getUri(docName),
+                range: new vscode.Range(line, 0, line, 0)
+              });
+            }
+          }
+        }
+      }
+      return result;
+    });
   }
 }